fix: correct ul style problem
Having defined custom css in the timeline.module.css the style
was inherited to the posts, causing the lists (both ordered and
unordered to end up with no style. I was going nuts with the bug,
as I didn't seem to find it.

Just adding the concrete styling to the timeline ul using the sx
prop fixed the problem.