feat: add start of cypress test config
I figured adding test would be nice addition to the
website. I went around trying to find a testing framework
that was powerful enough to give me all the testing functionality
to test the whole application.

I finally settled on cypress, which is an E2E testing framework,
with a nice how-to in the gatsby.js website.