feat: add initial cypress files
Besides that, I finally could run the yarn test:e2e script
which was before broken (in the previous commit) because for
some reason yarn didn't install date-fns, which is a dependency.

At the moment, everything is in place to start writing tests. There
is even example files. They will eventually be removed from the repo,
but I find having an initial files commit is not so bad.